Nix copy/pasted drawing code.
This code was copy/pasted from the SDL driver.  I'm not using it
so don't bother with it for now.

The SDL driver appears to only put updates on the screen when
sleep or getch is called as refreshing on each line is expensive.
I might have to do something like this as begin/end draw on each
line update is quite slow